Add files via upload
This is  the .py module that is running as a script. It takes a directory of pdf files and parses through the pdf's, searching for a specified list of key words, and returns a csv with a table containing the file name, the keywords that were found, and the sentences in which the words were found.. At the moment, the directory needs to be specified and the running list of keywords needs to be hardcoded in as well.